requirements.txt
appdirs==1.4.4
asgiref==3.7.2
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.12.2
black==23.12.1
blinker==1.7.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
consonance==0.1.2
cryptography==41.0.7
dissononce==0.34.3
Django==5.0
django-filter==23.5
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.14.0
exceptiongroup==1.1.0
Flask==3.0.0
h11==0.14.0
idna==3.4
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.5.1
MarkupSafe==2.1.3
MouseInfo==0.1.3
mypy==1.8.0
mypy-extensions==1.0.0
outcome==1.2.0
packaging==23.2
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.1.0
protobuf==4.25.1
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-axolotl==0.2.2
python-axolotl-curve25519==0.4.1.post2
python-dotenv==0.21.0
python3-xlib==0.15
pytweening==1.0.7
pytz==2023.3.post1
pywhatkit==5.4
requests==2.28.1
selenium==4.7.2
six==1.10.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.64.1
transitions==0.9.0
trio==0.22.0
trio-websocket==0.9.2
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
typing_extensions==4.9.0
urllib3==1.26.13
webdriver-manager==3.8.5
Werkzeug==3.0.1
wikipedia==1.4.0
wsproto==1.2.0
